requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
yfinance>=0.2.50
newsapi-python>=0.2.7
sentence-transformers>=3.0.0
qdrant-client>=1.12.0
openai>=1.0.0
vader-sentiment>=3.2.0
pandas>=2.0.0
numpy>=1.24.0
requests>=2.31.0